UAR-2691: Simplyfying BO structure for BIGlEntry and UAGlEntry and adding references to the Error Reporting Service
