+++ src/main/java/edu/arizona/kra/subaward/batch/InvoiceFeedUtils.java
@@ -1,13 +1,8 @@
 package edu.arizona.kra.subaward.batch;
 
-import edu.arizona.kra.subaward.batch.bo.BiGlEntry;
-import edu.arizona.kra.subaward.batch.bo.UAGlEntry;
-import edu.arizona.kra.subaward.batch.bo.UAGlEntryAdapter;
-
 import java.sql.Date;
-import java.util.ArrayList;
 import java.util.Calendar;
-import java.util.List;
+
 
 /**
  * Class holding utility methods for the Subaward Invoice Feed Batch jobs
@@ -20,17 +15,6 @@ public final class InvoiceFeedUtils {
         c.setTime(date);
         c.add(Calendar.DATE, -days);
         Date result = new Date(c.getTimeInMillis());
-        return result;
-    }
-
-
-    public static final List<UAGlEntry> translateBiGlEntries( List<BiGlEntry> biGlEntryList){
-        List<UAGlEntry> result = new ArrayList<>(biGlEntryList.size());
-        UAGlEntryAdapter adapter = new UAGlEntryAdapter();
-        for( BiGlEntry biGlEntry: biGlEntryList){
-            UAGlEntry uaGlEntry =  adapter.translate(biGlEntry);
-            result.add(uaGlEntry);
-        }
         return result;
     }
