+++ src/main/java/edu/arizona/kra/subaward/batch/bo/BiGlEntry.java
@@ -1,11 +1,11 @@
 package edu.arizona.kra.subaward.batch.bo;
 
 /**
- * Represents a 'simplified' GL Entry for UAR for the Subaward Invoice Feed
+ * Represents a 'simplified' GL Entry for UAR from the BI's view of the KFS.GL_ENTRY table
  *
  * Created by nataliac on 8/6/18.
  */
-public class BiGlEntry extends GlEntry {
+public class BiGlEntry extends UAGlEntry {
 
     public BiGlEntry() {
     }
