+++ src/main/java/edu/arizona/kra/subaward/batch/service/impl/SubawardInvoiceFeedServiceImpl.java
@@ -1,9 +1,11 @@
 package edu.arizona.kra.subaward.batch.service.impl;
 
 import edu.arizona.kra.subaward.batch.InvoiceFeedConstants;
+import edu.arizona.kra.subaward.batch.service.SubawardInvoiceErrorReportService;
 import edu.arizona.kra.subaward.batch.service.SubawardInvoiceFeedService;
 import org.apache.cxf.common.util.CollectionUtils;
 import org.apache.cxf.common.util.StringUtils;
+import org.kuali.kra.infrastructure.KraServiceLocator;
 import org.kuali.rice.coreservice.framework.parameter.ParameterService;
 
 import java.util.ArrayList;
@@ -20,6 +22,8 @@ public class SubawardInvoiceFeedServiceImpl implements SubawardInvoiceFeedServic
 
     List <String> runSchedule;
     List <Integer> dataIntervals;
+
+    private SubawardInvoiceErrorReportService subawardInvoiceErrorReportService;
 
     public boolean isSubawardInvoiceFeedEnabled(){
         if (parameterService.parameterExists(InvoiceFeedConstants.PARAM_NAMESPACE_SUBWAWARD, InvoiceFeedConstants.PARAM_COMPONENT_BATCH, InvoiceFeedConstants.PARAM_NAME_SUBAWARD_INVOICE_FEED_ENABLED )) {
@@ -56,13 +60,10 @@ public class SubawardInvoiceFeedServiceImpl implements SubawardInvoiceFeedServic
                             dataIntervals.add(Integer.parseInt(interval));
                         } catch (NumberFormatException e){
                             LOG.error("getSubwawardInvoiceFeedDataIntervalsDays: Could not parse integer from SUBAWARD_INVOICE_DATA_INTERVALS:"+paramValue+" SKIPPING...");
+                            subawardInvoiceErrorReportService.recordError("getSubwawardInvoiceFeedDataIntervalsDays: Could not parse integer from SUBAWARD_INVOICE_DATA_INTERVALS:"+paramValue+" SKIPPING...", e);
                         }
                     }
                 }
-                //dummy data!!!!!!!!!!!!!!!!!
-//                dataIntervals.add(new Integer(3));
-//                dataIntervals.add(new Integer(1));
-//                dataIntervals.add(new Integer(1));
             }
         }
         return dataIntervals;
@@ -70,6 +71,14 @@ public class SubawardInvoiceFeedServiceImpl implements SubawardInvoiceFeedServic
 
     public void setParameterService(ParameterService parameterService) {
         this.parameterService = parameterService;
+    }
+
+
+    public SubawardInvoiceErrorReportService getSubawardInvoiceErrorReportService() {
+        if ( subawardInvoiceErrorReportService ==null ) {
+            subawardInvoiceErrorReportService =   KraServiceLocator.getService(SubawardInvoiceErrorReportService.class);
+        }
+        return subawardInvoiceErrorReportService;
     }
