+++ src/main/resources/edu/arizona/kra/subaward/UASubawardSpringBeans.xml
@@ -8,6 +8,8 @@
   <import resource="classpath:edu/arizona/kra/CustomCoreSpringBeans.xml"/>
   <import resource="classpath:edu/arizona/kra/UA-BiDataSourceSpringBeans.xml"/>
 
+  <bean id="subawardInvoiceErrorReportService" class="edu.arizona.kra.subaward.batch.service.impl.SubawardInvoiceErrorReportServiceImpl"/>
+
   <bean id="subawardInvoiceFeedJob" parent="scheduledJobDescriptor" >
     <property name="steps">
       <list>
@@ -17,9 +19,8 @@
   </bean>
 
   <bean id="importGLDataStep" class="edu.arizona.kra.subaward.batch.ImportGLDataStep" parent="step">
-    <property name="glDataImportService">
-      <ref bean="glDataImportService"/>
-    </property>
+    <property name="glDataImportService" ref="glDataImportService"/>
+    <property name="subawardInvoiceErrorReportService" ref="subawardInvoiceErrorReportService"/>
   </bean>
 
 
@@ -27,17 +28,13 @@
     <property name="dataSource" ref="arizona.edu.biNonTransactionalDataSource"/>
   </bean>
 
-  <bean id="SubawardGlFeedDao" parent="platformAwareDao"
-        class="edu.arizona.kra.subaward.batch.dao.ojb.SubawardGlFeedDaoOjb"/>
+  <bean id="SubawardGlFeedDao" parent="platformAwareDao" class="edu.arizona.kra.subaward.batch.dao.ojb.SubawardGlFeedDaoOjb"/>
 
   <bean id="glDataImportService" class="edu.arizona.kra.subaward.batch.service.impl.GlDataImportServiceImpl">
     <property name="businessObjectService" ref="businessObjectService"/>
-    <property name="biGLFeedDao">
-      <ref bean="BiGlFeedDao"/>
-    </property>
-    <property name="subawardGLFeedDao">
-      <ref bean="SubawardGlFeedDao"/>
-    </property>
+    <property name="biGLFeedDao" ref="BiGlFeedDao"/>
+    <property name="subawardGLFeedDao" ref="SubawardGlFeedDao"/>
+    <property name="subawardInvoiceErrorReportService" ref="subawardInvoiceErrorReportService"/>
   </bean>
 
 </beans>
